+++ src/main/java/com/dunice/GoncharovVVAdvancedServer/repository/NewsRepository.java
@@ -3,5 +3,5 @@ package com.dunice.GoncharovVVAdvancedServer.repository;
 import com.dunice.GoncharovVVAdvancedServer.entity.NewsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<NewsEntity, Integer> {
+public interface NewsRepository extends JpaRepository<NewsEntity, Long> {
 }
